+HP, +Bullet, +Damage

+++ Assets/AtomicTest/AtomikValue/BulletEntityValueAPI.cs
@@ -15,18 +15,18 @@ namespace testAtomic
     public static class BulletEntityValueAPI
     {
         ///Keys
-        public const int BulletDamage = 24; // int
+        public const int BulletDamage = 24; // float
 
 
         ///Extensions
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static int GetBulletDamage(this IEntity obj) => obj.GetValue<int>(BulletDamage);
+        public static float GetBulletDamage(this IEntity obj) => obj.GetValue<float>(BulletDamage);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool TryGetBulletDamage(this IEntity obj, out int value) => obj.TryGetValue(BulletDamage, out value);
+        public static bool TryGetBulletDamage(this IEntity obj, out float value) => obj.TryGetValue(BulletDamage, out value);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool AddBulletDamage(this IEntity obj, int value) => obj.AddValue(BulletDamage, value);
+        public static bool AddBulletDamage(this IEntity obj, float value) => obj.AddValue(BulletDamage, value);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool HasBulletDamage(this IEntity obj) => obj.HasValue(BulletDamage);
@@ -35,6 +35,6 @@ namespace testAtomic
         public static bool DelBulletDamage(this IEntity obj) => obj.DelValue(BulletDamage);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void SetBulletDamage(this IEntity obj, int value) => obj.SetValue(BulletDamage, value);
+        public static void SetBulletDamage(this IEntity obj, float value) => obj.SetValue(BulletDamage, value);
     }
 }
